requirements.txt
fastapi
sqlalchemy
pydantic
uvicorn
python-jose
python-dotenv
pytest
